+++ src/main/java/be/baur/sdt/statements/ForEachStatement.java
@@ -28,7 +28,7 @@ public class ForEachStatement extends XPathStatement {
 	
 	@SuppressWarnings("rawtypes")
 	@Override
-	public void execute(TransformContext tracon, StatementContext stacon) throws TransformException {
+	public void execute(TransformContext traco, StatementContext staco) throws TransformException {
 		/*
 		 * Execution: create an XPath from the statement expression, set the variable
 		 * context and evaluate it to obtain a node-set. Execute the compound statement
@@ -40,25 +40,25 @@ public class ForEachStatement extends XPathStatement {
 
 		try {
 			SDAXPath xpath = new SDAXPath(getExpression());
-			xpath.setVariableContext(stacon);
-			List nodeset = xpath.selectNodes(stacon.getContextNode());
+			xpath.setVariableContext(staco);
+			List nodeset = xpath.selectNodes(staco.getContextNode());
 			/* 
 			 * Breaks if this expression does not return a List of nodes, must fix this later! 
 			 */
 			if (nodeset.isEmpty()) return; // nothing to do
 			
-			StatementContext comcon = stacon.newChild(); // compound statement context
-			comcon.setVariableValue(SDT.FUNCTIONS_NS_URI, "last", new Double(nodeset.size()));
+			StatementContext coco = staco.newChild(); // compound statement context
+			coco.setVariableValue(SDT.FUNCTIONS_NS_URI, "last", new Double(nodeset.size()));
 			
 			int position = 0;
 			for (Object current : nodeset) {
 			
-				++position; comcon.setContextNode(current);
-				comcon.setVariableValue(SDT.FUNCTIONS_NS_URI, "current", current);
-				comcon.setVariableValue(SDT.FUNCTIONS_NS_URI, "position", new Double(position));
+				++position; coco.setContextNode(current);
+				coco.setVariableValue(SDT.FUNCTIONS_NS_URI, "current", current);
+				coco.setVariableValue(SDT.FUNCTIONS_NS_URI, "position", new Double(position));
 				
 				for (Node statement : statements) {
-					((Statement) statement).execute(tracon, comcon);
+					((Statement) statement).execute(traco, coco);
 				}
 			}
 		
@@ -74,7 +74,7 @@ public class ForEachStatement extends XPathStatement {
 	 */
 	public DataNode toSDA() {
 		DataNode node = new DataNode(Statements.FOREACH.tag, getExpression());
-		for (Node statement : nodes()) // // add child statements
+		for (Node statement : nodes()) // add child statements
 			node.add(((Statement) statement).toSDA());
 		return node;
 	}