+++ src/test/java/test/TestTransform.java
@@ -25,6 +25,6 @@ public final class TestTransform {
 		DataNode node = tran.execute(c); 
 		SDA.format(w, node);
 		w.write(">>\n");
-
+		w.close();
 	}
 }